Change summary to say that this package is "Red Hat family"
It's confusing to see a package that says that it's for Red Hat installed
in a Fedora installation. For historical reasons, this package is named
"redhat-*", and it'd be complicated to change this [1]. By saying that this
is for the "Red Hat _family_" we also give another broader interpratation
to the package name.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1829430

The hyphenation is hard: "Red Hat" is spelled with a space, but we need
hyphens to connect the modifiers to it.